feat. add new alarms
