+++ src/kivyGUI/SmartBen.kv
@@ -57,3 +57,6 @@
             text: 'Close'
             size_hint_y: 0.25
             on_release: root.dismiss()
+
+<ImgBtn>:
+    size_hint: None, None
